server/requirements.txt
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
Django==5.1.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.8
image==1.5.33
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
numpy-stl==3.1.2
onnxruntime==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.28.0
PyMatting==1.1.12
python-utils==3.8.2
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
tifffile==2024.8.30
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
gunicorn==20.1.0